YARN-6888. Refactor AppLevelTimelineCollector such that RM does not have aggregator threads created. Contributed by Vrushali C.
(cherry picked from commit 20dd6d18b7787e67ef96f3b6b92ea8415a8650fd)
